Rename `search/database/SearchDatabaseApi.ts` to `search/search.ts`
