+++ client/src/main/scripts/components/search/VideoResultsList.spec.tsx
@@ -19,7 +19,7 @@
  */
 import * as renderer from "react-test-renderer";
 import { FoundationHarness } from "../../common/foundationTest";
-import { searchImpl, SearchWithData } from "./database/SearchDatabaseApi";
+import { searchImpl, SearchWithData } from "./search";
 // this data is generated by server/src/test/java/controllers/LucenePlayground.java
 import socialSecuritySearchResults from "./testData/socialSecuritySearchResults.json";
 import { VideoResultsList } from "./VideoResultsList";
